+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/buildsi/spliced/tree/main) (0.0.x)
+ - libabigail should also run abidiff (0.0.15)
  - add spack test as a predictor for spack installs (0.0.14)
  - abicompat is missing the binary as the first argument! (0.0.13)
  - Mock splice missing install of splice and replace (0.0.12)